# SubastasWeb/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Subastas",
        "description": "Documentación de la API de Subastas",
        "contact": {
            "email": "contacto@subastas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/subastas": {
            "get": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Obtener todas las subastas",
                "operationId": "9e5756191155cc53466a54fdf7206531",
                "responses": {
                    "200": {
                        "description": "Lista de subastas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Crear una nueva subasta",
                "operationId": "97f89b25cddf0cd0d6276702da232475",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "duracionMinutos",
                                    "fecha",
                                    "casa_remate_id",
                                    "rematador_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "duracionMinutos": {
                                        "type": "integer"
                                    },
                                    "activa": {
                                        "type": "boolean"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-06-01T14:00:00Z"
                                    },
                                    "casa_remate_id": {
                                        "type": "integer"
                                    },
                                    "rematador_id": {
                                        "type": "integer"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "videoId": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subasta creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/subastas/{id}": {
            "get": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Obtener una subasta por ID",
                "operationId": "e64135c72ea7e8b7f07170d4b14cab69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subasta encontrada"
                    },
                    "404": {
                        "description": "Subasta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Actualizar una subasta",
                "operationId": "5980b5b15bdfa8004ee31d5802d3492f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "duracionMinutos": {
                                        "type": "integer"
                                    },
                                    "activa": {
                                        "type": "boolean"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-06-01T14:00:00Z"
                                    },
                                    "casa_remate_id": {
                                        "type": "integer"
                                    },
                                    "rematador_id": {
                                        "type": "integer"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "videoId": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "loteIndex": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subasta actualizada"
                    },
                    "404": {
                        "description": "Subasta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Eliminar una subasta por ID",
                "operationId": "94a4b3909295b2e53e146229ffa7d6c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subasta eliminada correctamente"
                    },
                    "404": {
                        "description": "Subasta no encontrada"
                    }
                }
            }
        },
        "/api/subastas/{id}/lotes": {
            "post": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Asociar lotes a una subasta",
                "operationId": "8931c862071ed28dacf37560b8480e58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lotes"
                                ],
                                "properties": {
                                    "lotes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lotes asociados exitosamente"
                    },
                    "404": {
                        "description": "Subasta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/subastas/enviaMail": {
            "post": {
                "tags": [
                    "Subastas"
                ],
                "summary": "Envía una notificación por email",
                "description": "Envía un email de notificación con asunto y mensaje personalizados",
                "operationId": "7d63d54067a5d94e4d07337c9f5d4e7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "asunto",
                                    "mensaje"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Dirección de email del destinatario",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "asunto": {
                                        "description": "Asunto del email",
                                        "type": "string"
                                    },
                                    "mensaje": {
                                        "description": "Contenido del mensaje",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email enviado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}